+++ pkg/providers/dns/resource_akamai_dns_zone.go
@@ -185,6 +185,11 @@ func resourceDNSv2Zone(config dnsZoneResourceConfig) *schema.Resource {
 					},
 				},
 			},
+			"multi_provider_dnssec": {
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Description: "Enables multi-signer DNSSEC for the zone.",
+			},
 			"version_id": {
 				Type:     schema.TypeString,
 				Computed: true,
@@ -701,6 +706,14 @@ func populateDNSv2ZoneState(d *schema.ResourceData, zoneresp *dns.GetZoneRespons
 		return fmt.Errorf("%w: %s", tf.ErrValueSet, err.Error())
 	}
 
+	multiProviderDnssecEnabled := false
+	if zoneresp.MultiProviderDNSSEC != nil {
+		multiProviderDnssecEnabled = zoneresp.MultiProviderDNSSEC.Enabled
+	}
+	if err := d.Set("multi_provider_dnssec", multiProviderDnssecEnabled); err != nil {
+		return fmt.Errorf("%w: %s", tf.ErrValueSet, err.Error())
+	}
+
 	tsigListNew := make([]interface{}, 0)
 	if zoneresp.TSIGKey != nil {
 		tsigNew := map[string]interface{}{
@@ -800,6 +813,14 @@ func populateDNSv2ZoneObject(d *schema.ResourceData, zone *dns.ZoneCreate, logge
 		}
 	}
 
+	multiProviderDnssec, err := tf.GetBoolValue("multi_provider_dnssec", d)
+	if err != nil && !errors.Is(err, tf.ErrNotFound) {
+		return err
+	}
+	zone.MultiProviderDNSSEC = &dns.MultiProviderDNSSEC{
+		Enabled: multiProviderDnssec,
+	}
+
 	TSIGKey, err := tf.GetListValue("tsig_key", d)
 	if err != nil && !errors.Is(err, tf.ErrNotFound) {
 		if !errors.Is(err, tf.ErrNotFound) {
@@ -850,6 +871,10 @@ func checkDNSv2Zone(d tf.ResourceDataFetcher) error {
 	if err != nil && !errors.Is(err, tf.ErrNotFound) {
 		return err
 	}
+	multiProviderDnssec, err := tf.GetBoolValue("multi_provider_dnssec", d)
+	if err != nil && !errors.Is(err, tf.ErrNotFound) {
+		return err
+	}
 	ztype := strings.ToUpper(zoneType)
 	masters := mastersSet.List()
 	if ztype == "SECONDARY" && len(masters) == 0 {
@@ -866,6 +891,9 @@ func checkDNSv2Zone(d tf.ResourceDataFetcher) error {
 	}
 	if signandserve && ztype == "ALIAS" {
 		return fmt.Errorf("sign_and_serve is not valid in %s zone %s configuration", ztype, zone)
+	}
+	if multiProviderDnssec && !signandserve {
+		return fmt.Errorf("multi_provider_dnssec requires sign_and_serve to be true in zone %s configuration", zone)
 	}
 	if ztype != "SECONDARY" && len(tsig) > 0 {
 		return fmt.Errorf("tsig_key can not be populated in %s zone %s configuration", ztype, zone)